added new DateRangePicker, removed old
